Add files via upload
Just finished the ggplot try. There are two issues for improvement: the title of the plot cannot indicate whether the scale is raw or log transformed; the axis cannot be shown as the code which I cannot figure it out at present. Help needed. Hope it is helpful. I checked Stefan's code and found he didn't put the log transformation inside the function.